rename: NetworkLog -> NetworkApi
