+++ modules/network-api/src/main/java/com/whoz_in/network_api/NetworkApiApplication.java
@@ -1,13 +1,13 @@
-package com.whoz_in.network_log;
+package com.whoz_in.network_api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication(scanBasePackages = {"com.whoz_in"})
-public class NetworkLogApplication {
+public class NetworkApiApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(NetworkLogApplication.class, args);
+        SpringApplication.run(NetworkApiApplication.class, args);
     }
 
 }
