+++ modules/network-api/src/main/java/com/whoz_in/network_api/controller/IpAddressController.java
@@ -1,9 +1,8 @@
-package com.whoz_in.network_log.controller;
+package com.whoz_in.network_api.controller;
 
-import com.whoz_in.network_log.common.util.RequesterInfo;
+import com.whoz_in.network_api.common.util.RequesterInfo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
